// apps/web/src/lib/locations/types.ts
export type LocationDetailItem = {
  id: string;
  name: string;
  isActive: boolean;
  branchCode: string | null;
  designation: string | null;
  locality: string | null;
  category: string | null;
  phone: string | null;
  mobilePhone: string | null;
  fax: string | null;
  companyName: string | null;
  email: string | null;
  street: string | null;
  postalCode: string | null;
};